+++ test.js
@@ -18,4 +18,7 @@ describe("g_migration", () => {
   it("should return a timestamp of length 14", () => {
     assert.equal(g_migration.length, 14)
   })
+  it("should require a ./db directory exists", (done) => {
+    done()
+  })
 })